AI just needs traversing
integrated with the main game, the agents just need to traverse their
paths. May have to restrict the total number of NPC's based on lag.
would be nice to add to the loading screen the current state of the
load.
need to make loading asynchronous, it currently hangs in a couple of
spots