pillars fixed and events show correctly
replaced all pillars with a pillar cap and then in the scene manually
added pillars in the corners and then delete the caps at those points on
load.
The engineer events that are not needed show the fixed prefab and the
needed ones show the broken prefab. On completion of an event, it does
not swap to the fixed asset.... YET!